Fix reinterpret_cast security bug in SurfaceFlinger
Bug is due to a reinterpret_cast used when obtaining a sp<Layer>
from an sp<IBinder> passed from a client.
Without a checking mechanism, client could pass a malicious data packet.
This is a modified cherry-pick of a patch by Rob Carr that utilizes a map
to identify the appropriate layer based on the incoming IBinder token.

Original patch commit:

Author: Robert Carr <racarr@google.com>
Date:   Thu Apr 11 13:18:21 2019 -0700

SurfaceFlinger: Validate layers before casting.

Reinterpret casting random IBinder = no-fun. I first attempted
to use inheritance of "getInterfaceDescriptor" in Layer::Handle but
departing from "standard-layout" (e.g. using virtual methods) means that
downcasting with static/reinterpret_cast is no longer valid. Instead I opted
for the pattern the system-server uses of maintaing a map.

Now that we look up the handle in a map rather than casting IBinder
to Layer::Handle we need to make sure we have unique instances of the
handle. In general this is true but we weren't doing this in the
createWithSurfaceParent where we had an extra call to getHandle. Here
we both refactor createWithSurfaceParent so it works with the new
changes and also add protection for getHandle. We also fix an error
where the handle map was populated outside of lock.